+++ java/com/turbovnc/rfb/CMsgReader.java
@@ -1,4 +1,5 @@
-/* Copyright (C) 2012, 2017-2018, 2022 D. R. Commander.  All Rights Reserved.
+/* Copyright (C) 2012, 2017-2018, 2022, 2025 D. R. Commander.
+ *                                           All Rights Reserved.
  * Copyright 2016, 2018-2019 Pierre Ossman for Cendio AB
  * Copyright (C) 2002-2005 RealVNC Ltd.  All Rights Reserved.
  *
@@ -345,6 +346,9 @@ public class CMsgReader {
           break;
         case RFB.ENCODING_VMWARE_LED_STATE:
           readVMwareLEDState();
+          break;
+        case RFB.ENCODING_EXTENDED_MOUSE_BUTTONS:
+          handler.enableExtMouseButtons();
           break;
         default:
           readRect(new Rect(x, y, x + w, y + h), encoding);