+++ java/com/turbovnc/rfb/RFB.java
@@ -1,5 +1,5 @@
-/* Copyright (C) 2011-2012, 2015-2018, 2021-2022, 2024 D. R. Commander.
- *                                                     All Rights Reserved.
+/* Copyright (C) 2011-2012, 2015-2018, 2021-2022, 2024-2025
+ *           D. R. Commander.  All Rights Reserved.
  * Copyright 2009, 2011, 2019 Pierre Ossman for Cendio AB
  * Copyright (C) 2011-2012 Brian P. Hinz
  * Copyright (C) 2002-2005 RealVNC Ltd.  All Rights Reserved.
@@ -219,6 +219,8 @@ public final class RFB {
 
   public static final int ENCODING_TIGHT_WITHOUT_ZLIB      = -317;
 
+  public static final int ENCODING_EXTENDED_MOUSE_BUTTONS  = -316;
+
   public static final int ENCODING_CONTINUOUS_UPDATES      = -313;
   public static final int ENCODING_FENCE                   = -312;
 
@@ -341,6 +343,8 @@ public final class RFB {
   public static final int BUTTON5_MASK = 16;
   public static final int BUTTON6_MASK = 32;
   public static final int BUTTON7_MASK = 64;
+  public static final int BUTTON8_MASK = 128;
+  public static final int BUTTON9_MASK = 256;
 
   //***************************************************************************
   // GII